requirements.txt
streamlit
tensorflow  
numpy
matplotlib
pillow  